backend/requirements.txt
flask==3.0.3
vaderSentiment==3.3.2
flask-login==0.6.3
gunicorn==22.0.0
flask-cors==4.0.1
requests==2.31.0
Pillow==10.4.0